Add a rectangular grid layout for adding new widgets
This appends widgets into a grid that is largely rectangular and
automatically adds and removes rows as needed.

The simple mechanism has some quirks when adding and removing widgets
when it completes a rectangle, allowing you to grow wider rectangles
rather than keep it square. The behaviour is not intuitive there.