Use interfaces and const correctness for data classes
SharedProperties and TextureStore store data for the application. It is
important to limit which parts of the application are allowed to modify
that data. The main renderers are views into the data, and should under
no circumstances modify it.  Unfortunetly, QOpenGLTexture::bind() is not
a const function (nor is release) and the TextureStore cannot be const
in the renderers because of this.

Switching to interfaces for these classes and construction in main and
passing to mainwindow is for general best-practice reasons and possible
easier testing. They are constructed outside of MainWindow as they have
little to do with the rendering, they hold data, and they hold it
regardless of the main window.