Rework cubeintersections to return rotationmatrix
This lets us handle the rotation in the shader, and composite it with
other matrix operations more easily. Specifically it should help us put
the grid scaling in the correct spot more easily.

Gets rid of the quite ugly box class.

Changes to a single vertex attribute that is used to calculate both
vertex coordinates and texture coordinates.