Fix texture-destruction at shutdown.
Textures are owned by the texture-store, which lived until application
destruction. At that point the windows were closed and the
OpenGLContexts were all destroyed. Textures need to be destroyed in the
presence of a OpenGLContext.

Ownership is therefore moved back to MainWindow, and made unique to
MainWindow. MainWindow is destroyed before the child widgets, and so the
OpenGLContext(s) are alive.